+++ src/tensor/tensor.cpp
@@ -1031,7 +1031,7 @@ std::string mv::Tensor::getLogID() const
     return "Tensor:" + getName();
 }
 
-std::vector<float> mv::Tensor::computeNumericStrides() const
+std::vector<unsigned> mv::Tensor::computeNumericStrides() const
 {
     return getOrder().computeByteStrides(shape_, getDType().getSizeInBits() / 8);
 }
